chore/fix/feat: add some functionality and fix error

+++ app/src/main/java/com/example/myrouteoptimization/data/source/remote/response/OptimizeResponse.kt
@@ -32,7 +32,7 @@ data class OptimizeResponseData(
 	val numberOfVehicles: Int,
 
 	@field:SerializedName("total_distance")
-	val totalDistance: Int,
+	val totalDistance: Float,
 
 	@field:SerializedName("title")
 	val title: String,
